+++ cdflib/dataclasses.py
@@ -1,0 +1,61 @@
+from dataclasses import dataclass
+from typing import List, Optional, Union
+
+import numpy as np
+
+
+@dataclass
+class CDRInfo:
+    encoding: int
+    copyright_: str
+    version: str
+    majority: int
+    format_: bool
+    md5: bool
+    post25: bool
+
+
+@dataclass
+class GDRInfo:
+    first_zvariable: int
+    first_rvariable: int
+    first_adr: int
+    num_zvariables: int
+    num_rvariables: int
+    num_attributes: int
+    rvariables_num_dims: int
+    rvariables_dim_sizes: List[int]
+    eof: int
+    leapsecond_updated: Optional[int] = None
+
+
+@dataclass
+class AEDR:
+    entry: np.ndarray
+    data_type: int
+    num_elements: int
+    next_aedr: int
+    entry_num: int
+    num_strings: Optional[int] = None
+
+
+@dataclass
+class VDR:
+    data_type: int
+    section_type: int
+    next_vdr_location: int
+    variable_number: int
+    head_vxr: int
+    last_vxr: int
+    max_rec: int
+    name: str
+    num_dims: int
+    dim_sizes: List[int]
+    compression_bool: bool
+    compression_level: int
+    blocking_factor: int
+    dim_vary: Union[List[int], List[bool]]
+    record_vary: int
+    num_elements: int
+    sparse: int
+    pad: Optional[bool] = None